move middleware to its own file; cleanup
