+++ runtime/io/common/src/aws/smithy/kotlin/runtime/io/middleware/Phase.kt
@@ -20,7 +20,7 @@ class Phase<Request, Response> : Middleware<Request, Response> {
         Before, After
     }
 
-    private val middlewares = mutableListOf<Middleware<Request, Response>>()
+    private val middlewares = ArrayDeque<Middleware<Request, Response>>()
 
     /**
      * Insert [interceptor] in a specific order into the set of interceptors for this phase
@@ -31,24 +31,28 @@ class Phase<Request, Response> : Middleware<Request, Response> {
     }
 
     /**
+     * Insert a [transform] that only modifies the request of this phase
+     */
+    fun register(transform: ModifyRequest<Request>, order: Order = Order.After) {
+        register(ModifyRequestMiddleware(transform), order)
+    }
+
+    /**
+     * Insert a [transform] that only modifies the response of this phase
+     */
+    fun register(transform: ModifyResponse<Response>, order: Order = Order.After) {
+        register(ModifyResponseMiddleware(transform), order)
+    }
+
+    /**
      * Register a middleware in a specific order
      */
     fun register(middleware: Middleware<Request, Response>, order: Order = Order.After) {
         when (order) {
-            Order.Before -> middlewares.add(0, middleware)
-            Order.After -> middlewares.add(middleware)
+            Order.Before -> middlewares.addFirst(middleware)
+            Order.After -> middlewares.addLast(middleware)
         }
     }
-
-    /**
-     * Register an interceptor that only modifies the request of this phase
-     */
-    fun register(fn: ModifyRequest<Request>, order: Order = Order.After) { register(ModifyRequestMiddleware(fn), order) }
-
-    /**
-     * Register an interceptor that only modifies the response of this phase
-     */
-    fun register(fn: ModifyResponse<Response>, order: Order = Order.After) { register(ModifyResponseMiddleware(fn), order) }
 
     // runs all the registered interceptors for this phase
     override suspend fun <H : Handler<Request, Response>> handle(request: Request, next: H): Response {
